requirements.txt
# Core ML & Deep Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0

# NLP & Text Processing
spacy>=3.5.0
pydantic>=2.0.0

# Web Framework
streamlit>=1.28.0

# Configuration
python-dotenv>=1.0.0
requests>=2.31.0

# Testing & Quality
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Development (optional)
jupyter>=1.0.0
notebook>=7.0.0
